Rename dartboard module to artboard
    - update app references to the renamed game module
    - rename matching integration tests and relax wrapped copy assertion
